06_debugging_arize_phoenix/requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-mistralai
arize-phoenix
arize-phoenix-client
openinference-instrumentation-langchain
pypdf
dotenv
sentence-transformers